add_data: Read variables from environment instead of command-line
Instead of manually passing the needed variables as command-line arguments
to add_data.sh in python/Dockerfile, rely on these variables being already
in the environment, as defined either in .env file for Docker Compose,
or in the task definition of Amazon ECS.

Also add a quick environment variable check at the beginning of
add_data.sh to warn if any variable is empty.

This fixes "curl: (3) URL using bad/illegal format or missing URL" error
in "Creating PSRA Kibana Index Patterns" due to unquoted command-line
arguments in python/Dockerfile, causing KIBANA_ENDPOINT to be empty
when the optional ES_USER and ES_PASS are empty.