Fix error in "cd" command in get_git_lfs_pointers_of_csv_files()
I earlier changed the repo to include the owner, i.e. by changing
e.g. earthquake-scenarios to OpenDRR/earthquake-scenarios,
but forgot to do a basename so as to "cd" to the correct directory path,
causing the stack build run to fail prematurely.

Fixes a bug that I again carelessly introduced in
commit 399aa4ab63769d47dafb20305de6ddad06708bd8